requirements.txt
streamlit
openai>=1.0.0
pypdf
numpy
python-dotenv